TWINFACES-205 Rename @/entities/permissionGroup to @/entities/permission-group
