+++ src/entities/permission-group/api/hooks/useFetchById.ts
