+++ src/entities/permission-group/api/permission-group-api.ts
