+++ src/entities/permission-group/components/index.ts
