+++ src/entities/permission-group/components/permission-group-icon.tsx
