+++ src/entities/permission-group/components/resource-link/index.ts
