+++ src/entities/permission-group/components/resource-link/resource-link.tsx
