+++ src/entities/permission-group/components/resource-link/tooltip.tsx
