+++ src/entities/permission-group/libs/helpers.ts
