+++ src/entities/permission-group/libs/hooks/index.ts
