+++ src/entities/permission-group/libs/hooks/useSelectAdapter.ts
