+++ src/entities/permission-group/libs/index.ts
