+++ src/features/api-context-provider/api-context-provider.tsx
@@ -34,7 +34,7 @@ import {
 import {
   createPermissionGroupApi,
   PermissionGroupApi,
-} from "@/entities/permissionGroup";
+} from "@/entities/permission-group";
 import {
   createPipelineStepApi,
   PipelineStepApi,
